requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
litellm==1.51.0
google-generativeai==0.8.3
pillow==10.1.0
python-jose==3.3.0
aiofiles==23.2.1
python-dotenv==0.21.0
setuptools